+++ actions/about.ts
@@ -3,25 +3,19 @@
 import { revalidateTag } from 'next/cache';
 import { redirect } from 'next/navigation';
 
-import {
-  deleteCareerCompany,
-  deleteClub,
-  deleteFacility,
-  postCareerCompany,
-  postCareerStat,
-  postClub,
-  postFacility,
-  putCareerCompany,
-  putCareerDescription,
-  putCareerStat,
-  putClub,
-  putContact,
-  putDirections,
-  putFacility,
-  putGreetings,
-  putHistory,
-  putOverview,
-} from '@/apis/about';
+import { putContact } from '@/apis/v2/about/contact';
+import { putDirections } from '@/apis/v2/about/directions/[id]';
+import { postFacility } from '@/apis/v2/about/facilities';
+import { deleteFacility, putFacility } from '@/apis/v2/about/facilities/[id]';
+import { putFutureCareers } from '@/apis/v2/about/future-careers';
+import { postCareerCompany } from '@/apis/v2/about/future-careers/company';
+import { deleteCareerCompany, putCareerCompany } from '@/apis/v2/about/future-careers/company/[id]';
+import { postCareerStat, putCareerStat } from '@/apis/v2/about/future-careers/stats';
+import { putGreetings } from '@/apis/v2/about/greetings';
+import { putHistory } from '@/apis/v2/about/history';
+import { putOverview } from '@/apis/v2/about/overview';
+import { postClub, putClub } from '@/apis/v2/about/student-clubs';
+import { deleteClub } from '@/apis/v2/about/student-clubs/[id]';
 import { CareerStatEditorContent } from '@/components/editor/CareerStatEditor';
 import {
   FETCH_TAG_CAREER,
@@ -84,7 +78,7 @@ const careerPath = getPath(futureCareers);
 
 export const putCareerDescriptionAction = withErrorHandler(
   async (data: { koDescription: string; enDescription: string }) => {
-    await putCareerDescription(data);
+    await putFutureCareers(data);
     revalidateTag(FETCH_TAG_CAREER);
     redirect(careerPath);
   },
