+++ apis/v1/about/history.ts
@@ -1,0 +1,9 @@
+import { getRequest } from '@/apis';
+import { FETCH_TAG_HISTORY } from '@/constants/network';
+import { AboutContent } from '@/types/about';
+import { Language } from '@/types/language';
+
+export const getHistory = (language: Language) =>
+  getRequest<AboutContent>(`/v1/about/history?language=${language}`, undefined, {
+    next: { tags: [FETCH_TAG_HISTORY] },
+  });
