+++ apis/v2/about/contact.ts
@@ -1,0 +1,4 @@
+import { putRequest } from '@/apis';
+
+export const putContact = (formData: FormData) =>
+  putRequest('/v2/about/contact', { body: formData, jsessionID: true });
