+++ apis/v2/about/directions/index.ts
@@ -1,0 +1,9 @@
+import { getRequest } from '@/apis';
+import { FETCH_TAG_DIRECTIONS } from '@/constants/network';
+import { Direction } from '@/types/about';
+import { WithLanguage } from '@/types/language';
+
+export const getDirections = () =>
+  getRequest<WithLanguage<Direction>[]>('/v2/about/directions', undefined, {
+    next: { tags: [FETCH_TAG_DIRECTIONS] },
+  });
