+++ apis/v2/about/facilities/index.ts
@@ -1,0 +1,12 @@
+import { getRequest, postRequest } from '@/apis';
+import { FETCH_TAG_FACILITIES } from '@/constants/network';
+import { Facility } from '@/types/about';
+import { WithLanguage } from '@/types/language';
+
+export const getFacilities = () =>
+  getRequest<WithLanguage<Facility>[]>('/v2/about/facilities', undefined, {
+    next: { tags: [FETCH_TAG_FACILITIES] },
+  });
+
+export const postFacility = (formData: FormData) =>
+  postRequest('/v2/about/facilities', { body: formData, jsessionID: true });
