+++ apis/v2/about/greetings.ts
@@ -1,0 +1,4 @@
+import { putRequest } from '@/apis';
+
+export const putGreetings = (formData: FormData) =>
+  putRequest('/v2/about/greetings', { body: formData, jsessionID: true });
