+++ apis/v2/about/history.ts
@@ -1,0 +1,4 @@
+import { putRequest } from '@/apis';
+
+export const putHistory = (formData: FormData) =>
+  putRequest('/v2/about/history', { body: formData, jsessionID: true });
