+++ apis/v2/about/overview.ts
@@ -1,0 +1,4 @@
+import { putRequest } from '@/apis';
+
+export const putOverview = (formData: FormData) =>
+  putRequest('/v2/about/overview', { body: formData, jsessionID: true });
