+++ apis/v2/about/student-clubs/[id].ts
@@ -1,0 +1,4 @@
+import { deleteRequest } from '@/apis';
+
+export const deleteClub = (id: number) =>
+  deleteRequest(`/v2/about/student-clubs/${id}`, { jsessionID: true });
