+++ app/[locale]/about/future-careers/page.tsx
@@ -1,6 +1,6 @@
 export const dynamic = 'force-dynamic';
 
-import { getFutureCareeres } from '@/apis/about';
+import { getFutureCareeres } from '@/apis/v1/about/future-careers';
 import { EditButton } from '@/components/common/Buttons';
 import LoginVisible from '@/components/common/LoginVisible';
 import HTMLViewer from '@/components/editor/HTMLViewer';
